Clean up code for reading initial values of global arrays and structures

+++ src/boomerang-gui/Decompiler.cpp
@@ -332,7 +332,7 @@ void Decompiler::getCompoundMembers(const QString& name, QTableWidget *tbl)
 
     std::shared_ptr<CompoundType> c = ty->as<CompoundType>();
 
-    for (size_t i = 0; i < c->getNumMembers(); i++) {
+    for (int i = 0; i < c->getNumMembers(); i++) {
         tbl->setRowCount(tbl->rowCount() + 1);
         tbl->setItem(tbl->rowCount() - 1, 0, new QTableWidgetItem(tr("%1").arg(c->getMemberOffsetByIdx(i))));
         tbl->setItem(tbl->rowCount() - 1, 1, new QTableWidgetItem(tr("%1").arg(c->getMemberOffsetByIdx(i) / 8)));
